+++ src/main/scala/exercise/cashier/Item.scala
@@ -6,14 +6,14 @@ object Item {
 
   val separator = '，'
 
-  def apply(product: Product, quantity: Int) = {
-    new Item(product, quantity, quantity * product.unitPrice, 0, None)
+  def apply(product: Product, quantity: BigDecimal) = {
+    new Item(product, quantity, (quantity * BigDecimal(product.unitPrice)).intValue(), 0, None)
   }
 }
 
 case class Item(
     product: Product,
-    quantity: Int,
+    quantity: BigDecimal,
     amount: Int,
     saved: Int,
     promotion: Option[Promotion]) {
